Event when input size point change Event when input size segment change Event when input size curve change Event when input size texte change Event when input color background change Event when input color point change Event when input color segment change Event when input color curve change Event when input color texte change Event when use save settings button Copie user image in www/images/tmp so javascript can handle it

Event when input size point change Event when input size segment change Event when input size curve change Event when input size texte change Event when input color background change Event when input color point change Event when input color segment change Event when input color curve change Event when input color texte change Event when use save settings button Copie user image in www/images/tmp so javascript can handle it

server(input, output, session)

Arguments

input
: Input shiny

images
dataframe : info image (name, path) Event when use load button Event when use next button Event when use save button Event when use previous button Event when use first button Event when use last button Action when user change selected image Get the os of MoMacs run

yaml_file
: Path of yml file Save yml file

images
:

result
: the path of yml file Save mom file

images
:

input
: Input shiny

result
: the path of mom file Event end image Event when use Save Shortcut button

Value

string End event Event when use shinyFilesButton Event when use shinyDirButton Load yml file